Session variables and env abailabe
The API is now able to generate an session_id and store it at the client.
(This will change in the future with the implementation of JWT.)

Also, a new package has been added to the back-end API, that whay
the back-end has access to the new .env file variables.