Login implemented
At the front-end a middleware is set up to keep track of cookies,
when a token is set up in a cookie, and it is valid, the page will
redirect the user to the wanted route, if the token aren't set or
isn't valid, the user will be sent to the login page.

A new token can be set in a cookie when the user login correctly.